Improve handling of two-digit DOB years
This caused a slightly silly bug in clinic where patients born after
2000 would generate stickers where they'd be over a century old. This
is because QFlow only exports DOBs with years as two-digits :(

We 'fix' this by prefixing with 20- instead of -19 on a rolling basis
based on the current year. We have a little bit of leeway as it doesn't
look like we'll be vaccinating under-5s, so this should work well, but
it isn't a huge deal as errors in the DOB of this degree are very easily
spotted by personnel, and then the details can be manually entered.